func-apps/chat-summary/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1732732445543/work
annotated-types==0.7.0
anyio==4.8.0
archspec @ file:///croot/archspec_1709217642129/work
attrs==25.1.0
azure-functions==1.21.3
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1734380492396/work/certifi
cffi @ file:///croot/cffi_1726856441404/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1738241127085/work
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1733258728466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1731366181659/work
cryptography @ file:///croot/cryptography_1732129644106/work
distro @ file:///croot/distro_1714488253808/work
fastapi==0.115.8
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna @ file:///croot/idna_1714398848350/work
jiter==0.8.2
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
libmambapy @ file:///croot/mamba-split_1732896306512/work/libmambapy
menuinst @ file:///croot/menuinst_1731364904452/work
multidict==6.1.0
openai==0.28.0
packaging @ file:///croot/packaging_1720101850331/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///croot/pluggy_1733169602837/work
propcache==0.2.1
psycopg2-binary==2.9.10
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic_core==2.27.2
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv_1733243180666/work
PyYAML==6.0.2
requests @ file:///croot/requests_1730999120400/work
ruamel.yaml @ file:///croot/ruamel.yaml_1727980156079/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1727769819935/work
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tqdm @ file:///croot/tqdm_1724853939799/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
typing_extensions==4.12.2
urllib3 @ file:///croot/urllib3_1727769808118/work
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
wheel==0.44.0
yarl==1.18.3
zstandard @ file:///croot/zstandard_1731356346222/work
pydantic==1.9.0
pydantic-core==1.9.0
